First part of trigger refactor - Removed enum column on object_tag - Refactored object_tag trigger so it won't need to be updated if a new "taggable" table is created - Removed onDelete object triggers and added an @afterDelete() - Removed set controller validator and accept all strings - Removed all tag (jsonb) columns and replaced with a many to many relationship with tags - Changed most endpoints to return all the nested object/tags, and added order by - Updated Events, Event Types, Person Types to show tags component on individual form - Updated tags in tables to show a list of tag strings. (Currently no ellipse/row limit)
Merged Develop in
