+++ packages/admin/src/app/feature/project-management/all-people/all-people.component.html
@@ -43,7 +43,15 @@
   <ng-container matColumnDef="tags">
     <th mat-header-cell *matHeaderCellDef>Tags</th>
     <td mat-cell *matCellDef="let element">
-      {{ element.tags ? element.tags.length : 0 }}
+      @if (element.tags && element.tags.length > 0) {
+        @for (tag of element.tags; track $index) {
+          <div>
+            {{ tag.pathname }}
+          </div>
+        }
+      } @else {
+        <div>--</div>
+      }
     </td>
   </ng-container>
