+++ packages/admin/src/app/feature/project-management/person-type/person-type.component.scss
