+++ packages/admin/src/app/feature/project-management/person-types/person-types.component.html
@@ -45,7 +45,15 @@
   <ng-container matColumnDef="tags">
     <th mat-header-cell *matHeaderCellDef>Tags</th>
     <td mat-cell *matCellDef="let element">
-      <pre>{{ element.tags ? (element.tags | json) : '' }}</pre>
+      @if (element.tags && element.tags.length > 0) {
+        @for (tag of element.tags; track $index) {
+          <div>
+            {{ tag.pathname }}
+          </div>
+        }
+      } @else {
+        <div>--</div>
+      }
     </td>
   </ng-container>
