+++ packages/admin/src/app/feature/project-management/project-management.routes.ts
@@ -6,7 +6,7 @@ import { PlaceholderComponent } from '@feature/project-management/placeholder/pl
 import { ProjectManagementComponent } from './project-management.component';
 import { AllPeopleComponent } from '@feature/project-management/all-people/all-people.component';
 import { PersonComponent } from '@feature/project-management/person/person.component';
-import { PersonTypeComponent } from '@feature/project-management/person_type/person-type.component';
+import { PersonTypeComponent } from '@feature/project-management/person-type/person-type.component';
 import { EventComponent } from '@feature/project-management/event/event.component';
 import { AllEventTypesComponent } from './all-event-types/all-event-types.component';
 import { EventTypeComponent } from './event-type/event-type.component';
